First commit.
This code/interface worked in a different filesystem layout. In an
attempt to improve the state of affairs before implementing support for
multiple operating systems, things were moved around (and, in some
cases, renamed) so that the whole project "made sense."

Previous repositories (history not included) included a great deal of
cruft that has nothing to do with the "final" state of the server.

Changes will be coming to make this code run.