These changes let me build a clickable Mac app.
They should be compatible with building on other platforms as well. By
using the PLT GUI toolkit, I spawn an interface (first) for launching
the server, use that to set some critical parameters, and then the user
plays with the browser-based interface.

This was, put simply, what I wanted... double-click to launch, and then
proceed to play.