This seems to make Linux functional.
This was tested on a 10.x series Ubuntu. It may not work under other flavors, and requires testing.

However, this will compile, run, and upload to an Uno. It will also install the firmware. I have not yet tested the executable version, which will come next. However,

racket gui.rkt

launches a functional server and opens a browser.