Updating... implemented?
It looks like I have auto-update implemented.

It pulls the server.rkt file from GitHub, checks all of the versions in
the versions list, and if the local version is less than the remote
version, it pulls that file.

It also pulls the server.rkt file, so that the next time through we
don't re-update everything again.

I can also update flow.module, and possibly other files in the future.
However, that will require a rewrite of lists to hash tables, as we need
a way of growing the list of files that is being checked. Currently, I
do a for-each, which means the lists have to be the same size... which
does not work if I add a new file to versioning.

I do not currently have a way to update the core application...
although, I suppose I could do that, too?