Attempting to implement Windows support...
This may have broken things on Mac/Linux. Or, it may have broken nothing
on the other platforms.

I'm currently struggling with getting occ21 to eat some library search
paths. I have no idea if this is a "space in filename" issue, or if it
is something else entirely. However, this is annoying, and I really
don't like Windows.