Fixed for compilation.
It looks like I need a different path for command-line execution and
compilation. At this point, I'll take the hit and flip the UMBRELLA
variable so that things work under Windows for compilation.

More exploration there is necessary... the differences in how paths are
found (depending on executable used) is really annoying.