Adds error handling for missing internets, upload.
Handles the case when you are not connected to the internet (WAT?), and
also adds an upload to the build script for Mac. Will do the same for
Windows and Linux.

I want to be able to update the download package, but be able to see how
many times we get downloads. This will allow me to look at the httpd
logs and get a sense for how many people are exploring things. I could
do the same with GitHub, but it's annoying to automate the upload for
them.