This adds sort order to block sets.
This is all kludgy now. There's a good deal of work that now needs to be
done to make all of this prettier. Maybe the magic of opensource will
lead someone to fix everything.

Most likely, I'll have to do that.

There's a number of places (updating files remotely, etc.) where I could
have more abstraction/reuse, which would condense the code and make it
simpler. Likewise, the config file needs to be rethought.