Updating avrdude for all platforms.
The Mac and Windows platforms were updated from 5.4 to 5.11 in the official distribution. This made a difference on some Freeduinos the students built... but not all of them. In other words, the old version worked fine with my Freeduino, worked on one or two student boards... and then, not at all on others.

Weird. This should fix that problem.