Adding light sensing module.
Also, fixed the server. When printing to the text area, you can't have too much stuff in it, or it borks. When it does, Racket does not cleanup nicely, and we hold on to a serial port... which is bad.

I reset the string length after every 300 lines, so that should take care of it until I find a better solution.